dockerization/requirements.txt
tensorflow-cpu==2.1.0
Flask==1.1.1
tensorflow-text==2.1.1
